fix(sql): don't drop or rename the old prediction history table
Dropping prediction history caused dead lock when deploying to test
environment. Solution was to keep the original history table and create
a new history table side by side with the old one.

NOTE: The old history table can be removed in future releases when it is not
in use anymore.

BREAKING CHANGES:
This commit modifies an existing db migration script (V13_1) and is incompatible
with versions including this script.