feat(history): save facility status history
Facility status is saved every time the status is changed. The status
history is saved to a new table named facility_status_history. Saved
fields include facilityId, status, and status descriptions as well as
start and end dates. End date is null for the latest entry.

Initial statuses are not populated for pre-existing facilities, i.e.,
the first status history line will be generated when the facilitiy is
updated the next time.